+++ multi-window-demo/mwdemo/layer.hpp
@@ -16,10 +16,10 @@ class MWExampleLayer final : public Layer
     MWExampleLayer(IMultiWindowApplication *p_Application) noexcept;
 
     void OnStart() noexcept override;
-    void OnUpdate(usize p_WindowIndex) noexcept override;
-    void OnRender(usize p_WindowIndex, VkCommandBuffer p_CommandBuffer) noexcept override;
+    void OnUpdate(u32 p_WindowIndex) noexcept override;
+    void OnRender(u32 p_WindowIndex, VkCommandBuffer p_CommandBuffer) noexcept override;
     void OnImGuiRender() noexcept override;
-    bool OnEvent(usize p_WindowIndex, const Event &p_Event) noexcept override;
+    bool OnEvent(u32 p_WindowIndex, const Event &p_Event) noexcept override;
 
   private:
     IMultiWindowApplication *m_Application = nullptr;
