+++ onyx/onyx/rendering/render_specs.cpp
@@ -22,10 +22,10 @@ void ApplyCoordinateSystemIntrinsic(fmat4 &p_Transform) noexcept
 namespace Onyx::Detail
 {
 template <Dimension D, DrawMode DMode>
-PolygonDeviceInstanceData<D, DMode>::PolygonDeviceInstanceData(const usize p_Capacity) noexcept
+PolygonDeviceInstanceData<D, DMode>::PolygonDeviceInstanceData(const u32 p_Capacity) noexcept
     : DeviceInstanceData<InstanceData<D, DMode>>(p_Capacity)
 {
-    for (usize i = 0; i < ONYX_MAX_FRAMES_IN_FLIGHT; ++i)
+    for (u32 i = 0; i < ONYX_MAX_FRAMES_IN_FLIGHT; ++i)
     {
         VertexBuffers[i] = Core::CreateMutableVertexBuffer<D>(p_Capacity);
         IndexBuffers[i] = Core::CreateMutableIndexBuffer(p_Capacity);
@@ -33,7 +33,7 @@ PolygonDeviceInstanceData<D, DMode>::PolygonDeviceInstanceData(const usize p_Cap
 }
 template <Dimension D, DrawMode DMode> PolygonDeviceInstanceData<D, DMode>::~PolygonDeviceInstanceData() noexcept
 {
-    for (usize i = 0; i < ONYX_MAX_FRAMES_IN_FLIGHT; ++i)
+    for (u32 i = 0; i < ONYX_MAX_FRAMES_IN_FLIGHT; ++i)
     {
         VertexBuffers[i].Destroy();
         IndexBuffers[i].Destroy();
